requirements.txt
Pillow~=8.3.1
PyYAML~=5.4.1
base58~=2.1.0
bson~=0.5.10
python-escpos~=3.0a8
